Revert "rcar-vin: rcar-csi2: Add blank margin when caluculating bit rate"
This reverts commit 01d619a392ac781da05c4622abcd73b8a76df982.

The reason is follows.
When pixel clock rate with Hblank/Vblank margin added from decoder
side was propageted, this calculation is unnecessary and should be
removed. Also, it is not reasonable to calculate the margin with
the CSI2 driver. The propagating the pixel clock rate that margined
Hblank/Vblank on the decoder driver side is correct.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>