arm64: dts: renesas: salvator: add a connection from adv748x codec (HDMI input) to the R-Car SoC
Not sure if all variants of the Salvator board have the HDMI decoder
chip (the ADV7482) connected to the SSI4 on R-Car SoC, as it is on
Salvator-X ES1, so the the ADV7482 endpoint and connection definitions
are placed in the board file.

The I2C wiring is more likely to persist across the variants (similar
to ak4613, connected to the same interface), so that is in the common
file.