arm64: config+dts: cetibox: Enable USB Peripheral support
Align the cetibox VC2 device tree USB3 setup with the way its done in the
Salvator-X device tree. That way the USB3 xHCI and Peripheral controllers
are set up correctly.
Add a "companion" attribute to the USB3 Peripheral controller to enable the
usb-role-switch device.

Enable the R-Car Gen3 USB3 PHY driver and various USB gadget options.

With this setup, the CETiBox VC2 was tested succesfully as a USB device
(using the mass storage gadget) with both USB2 and USB3 hosts.